Update sv_miner.lua
try this code as it might fix your problem. it has a check function to help stop that issue from happening. if this helps or doesnt fix anything let me know and we can work on it. I apologize if the translations arent accurate as i used google translate.

hope this helped